+++ tests/testthat/test-print.R
@@ -190,8 +190,7 @@ test_that("comparison results can be printed", {
 })
 
 test_that("permutation test results can be printed for clusters", {
-  model <- group_tna(engagement_mmm)
-  perm <- permutation_test(model, iter = 50)
+  perm <- permutation_test(mmm_model, iter = 50)
   expect_error(
     capture.output(print(perm)),
     NA